refactor: move calculateCartTotals to utils and update imports
